# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Classroom Reservation API",
    "version": "1.0.0",
    "description": "API for managing classroom reservations"
  },
  "servers": [
    {
      "url": "http://localhost:5000"
    }
  ],
  "paths": {
    "/activities": {
      "get": {
        "summary": "获取所有活动",
        "description": "返回所有活动的列表",
        "tags": [
          "Activities"
        ],
        "responses": {
          "200": {
            "description": "成功获取活动列表",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Activity"
                  }
                }
              }
            }
          },
          "500": {
            "description": "服务器错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "创建活动",
        "description": "创建一个新的活动",
        "tags": [
          "Activities"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActivityInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "活动创建成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Activity"
                }
              }
            }
          },
          "400": {
            "description": "无效的请求数据",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "服务器错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/activities/{id}": {
      "get": {
        "summary": "根据 ID 获取活动",
        "description": "返回指定 ID 的活动",
        "tags": [
          "Activities"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "活动 ID"
          }
        ],
        "responses": {
          "200": {
            "description": "成功获取活动",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Activity"
                }
              }
            }
          },
          "404": {
            "description": "活动未找到",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "服务器错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "更新活动",
        "description": "更新指定 ID 的活动",
        "tags": [
          "Activities"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "活动 ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActivityInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "活动更新成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Activity"
                }
              }
            }
          },
          "400": {
            "description": "无效的请求数据",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "活动未找到",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "服务器错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "删除活动",
        "description": "删除指定 ID 的活动",
        "tags": [
          "Activities"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "活动 ID"
          }
        ],
        "responses": {
          "200": {
            "description": "活动删除成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "活动删除成功"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "活动未找到",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "服务器错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/classes": {
      "get": {
        "summary": "获取所有班级",
        "description": "返回所有班级的列表",
        "tags": [
          "Classes"
        ],
        "responses": {
          "200": {
            "description": "成功获取班级列表",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Class"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "创建班级",
        "description": "创建一个新的班级",
        "tags": [
          "Classes"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClassInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "班级创建成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Class"
                }
              }
            }
          },
          "400": {
            "description": "无效的输入"
          }
        }
      }
    },
    "/classes/{id}": {
      "get": {
        "summary": "根据 ID 获取班级",
        "description": "返回指定 ID 的班级",
        "tags": [
          "Classes"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "班级 ID"
          }
        ],
        "responses": {
          "200": {
            "description": "成功获取班级",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Class"
                }
              }
            }
          },
          "404": {
            "description": "班级未找到"
          }
        }
      },
      "put": {
        "summary": "更新班级",
        "description": "更新指定 ID 的班级",
        "tags": [
          "Classes"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "班级 ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClassInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "班级更新成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Class"
                }
              }
            }
          },
          "400": {
            "description": "无效的输入"
          },
          "404": {
            "description": "班级未找到"
          }
        }
      },
      "delete": {
        "summary": "删除班级",
        "description": "删除指定 ID 的班级",
        "tags": [
          "Classes"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "班级 ID"
          }
        ],
        "responses": {
          "200": {
            "description": "班级删除成功"
          },
          "404": {
            "description": "班级未找到"
          }
        }
      }
    },
...
  },
...
}